# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Social Hub API
  version: "0.1.0"
  description: >
    Unified Social Media Automation & Customer Messaging Platform.
    Auth uses an HttpOnly session cookie (see /api/auth/login) — there is no
    bearer token for the dashboard's own API; webhook endpoints instead use
    per-platform signature verification (see INTEGRATIONS.md).
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - cookieAuth: []
paths:
  /api/auth/login:
    post:
      summary: Log in and receive a session cookie
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
      responses:
        "200": { description: Logged in, session cookie set }
        "401": { description: Invalid credentials }
...
  /api/customers:
    get:
      summary: Search/list customers
      parameters:
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200": { description: Customer list }
  /api/integrations:
    get:
      summary: List connection status for every platform adapter
      responses:
        "200": { description: Integration rows + current APP_MODE }
    post:
      summary: Connect a platform (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform]
              properties:
                platform: { type: string, enum: [whatsapp, messenger, instagram, tiktok] }
      responses:
        "200": { description: Updated integration row }
        "403": { description: Admin only }
  /api/integrations/disconnect:
    post:
      summary: Disconnect a platform (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform]
              properties:
                platform: { type: string }
      responses:
        "200": { description: OK }
  /api/automations:
    get:
      summary: List automations
      responses:
        "200": { description: Automation list }
    post:
      summary: Create an automation rule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AutomationInput" }
      responses:
        "201": { description: Created }
        "403": { description: Missing MANAGE_AUTOMATION permission }
  /api/employees:
    get:
      summary: List employees in the organization
      responses:
        "200": { description: Employee list }
    post:
      summary: Create an employee (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                role: { type: string, enum: [ADMIN, EMPLOYEE] }
                permissions:
                  type: array
                  items: { type: string }
      responses:
        "201": { description: Created }
        "403": { description: Admin only }
...
  /api/notifications:
    get:
      summary: List notifications (Section 26)
      parameters:
        - in: query
          name: unread
          schema: { type: string, enum: ["true"] }
      responses:
        "200": { description: "{ notifications[], unreadCount }" }
    patch:
      summary: Mark one or all notifications as read
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string }
                all: { type: boolean }
      responses:
        "200": { description: OK }
...
  /api/ai/extract-intent:
    post:
      summary: Extract structured customer intent from a conversation's latest message (Section 23)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conversationId]
              properties:
                conversationId: { type: string }
      responses:
        "200": { description: "{ intent, product, questions[], confident }" }
        "400": { description: AI disabled }
  /api/ai/suggest-reply:
    post:
      summary: Ask the optional AI module to suggest a reply for a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conversationId]
              properties:
                conversationId: { type: string }
      responses:
        "200":
          description: Suggested reply, or confident=false with a human-handoff fallback message
        "400": { description: AI disabled (no OPENAI_API_KEY / AI_ENABLED=false) }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: social_hub_session
...
  schemas:
...
    AutomationInput:
      type: object
      required: [name, conditions, actions]
      properties:
        name: { type: string }
        triggerType: { type: string, enum: [new_message], default: new_message }
        conditions:
          type: object
          description: 'e.g. { "contains": "price" } or { "isNewCustomer": true }'
        actions:
          type: array
          items:
            type: object
            description: 'e.g. { "type": "send_reply", "text": "..." } | { "type": "add_tag", "tagName": "..." } | { "type": "assign_employee", "employeeId": "..." } | { "type": "notify_employee" }'
        enabled: { type: boolean, default: true }